OptionListContainer rename
I renamed the old OptionEditList to OptionListContainer to be have a more consistent naming scheme across the three different list classes. Other than renaming the files and class, I kept changes to a minimum (just one small adjustment in DtgeSceneEditContainer) to keep this specific commit clean.
